Add initial MD5 check logic
The idea is to extract the attachments to a temp directory and check
if the MD5 hashes are already known. At this moment, only the MD5 hashes
are checked against the files already extracted for one mail folder.
This needs to become a recursive process, where the files are moved and
checked again to the root folder of a user.

This still needs improvement, but this method seems to work reasonably
well with and at least performs better than calcualting hashes repeatanbly.